small fix for the sftp folder to work

+++ src/Helper/G4NTrait.php
@@ -413,12 +413,15 @@ trait G4NTrait
     public function makeTempFiles(array $files,  $filesystem):array
     {
         //with this we clear our temp files folder
-        $it = new RecursiveDirectoryIterator("uploads/temp", RecursiveDirectoryIterator::SKIP_DOTS);
-        $content = new RecursiveIteratorIterator($it,
-            RecursiveIteratorIterator::CHILD_FIRST);
-        foreach($content as $file) {
-            unlink($file->getRealPath());
+        if(file_exists('"uploads/temp"')) {
+            $it = new RecursiveDirectoryIterator("uploads/temp", RecursiveDirectoryIterator::SKIP_DOTS);
+            $content = new RecursiveIteratorIterator($it,
+                RecursiveIteratorIterator::CHILD_FIRST);
+            foreach ($content as $file) {
+                unlink($file->getRealPath());
+            }
         }
+        else mkdir('"uploads/temp"');
         $return = [];
         foreach ($files as $key => $file){
             $tempFile = 'temp/temp'.random_int(0, 10000).'.png';
